requirements.txt
aiofiles==0.7.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-CDN==1.5.3
Flask-CDN-NG==1.3.0
Flask-Login==0.5.0
Flask-Login-Multi==0.1.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.1
priority==2.0.0
pycparser==2.20
pymongo==3.11.4
pyOpenSSL==20.0.1
Quart==0.15.1
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wsproto==1.0.0
